refactor(createlocalvue): move createLocalVue to shared utils
